pyproject.toml
[tool.poetry]
name = "Example"
version = "0.1.0"
description = "This is example project"
authors = ["Vu Tran <vu.trandang@amela.vn>"]

[tool.poetry.dependencies]
python = "3.8.8"
Django = "^4.1"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
django-filter = "^22.1"
django-model-utils = "^4.2.0"
flake8 = "^5.0.4"
flake8-django = "^1.1.5"
psycopg2 = "^2.9.3"
pytest = "^7.1.2"
openpyxl = "^3.0.10"
munch = "^2.5.0"
PyYAML = "^6.0"
mixer = "^7.2.2"
PyHamcrest = "^2.0.4"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
